#13 Explanatory comments added to code
