+++ src/components/user-search-input/user-search-input.js
@@ -5,20 +5,25 @@ import AccountCircle from "@mui/icons-material/AccountCircle";
 import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
 import LoadingButton from "@mui/lab/LoadingButton";
 
+// UserSearchInput provides a TextField for username input.
 export function UserSearchInput({ onSubmit }) {
+  // State variables have been declared using destructuring.
   const [inputValue, setInputValue] = useState("");
   const [submitting, setSubmitting] = useState(false);
 
   const handleSubmit = async (event) => {
+    // preventDefault allows React to override the default action when a form is submitted.
     event.preventDefault();
     setSubmitting(true);
 
+    // Await a response (i.e. promise) from the onSubmit function in App.js before proceeding.
     await onSubmit(inputValue);
 
     setSubmitting(false);
     setInputValue("");
   };
 
+  // TextField is a controlled input, handleChange is called each time the input changes (i.e. as the user types).
   const handleChange = (event) => {
     const value = event.target.value;
     setInputValue(value);
@@ -29,6 +34,7 @@ export function UserSearchInput({ onSubmit }) {
       onSubmit={handleSubmit}
       style={{ display: "flex", alignItems: "center" }}
     >
+      {/* TextField is disabled when the form is submitting to prevent further input. */}
       <TextField
         fullWidth
         required
@@ -46,6 +52,7 @@ export function UserSearchInput({ onSubmit }) {
         disabled={submitting}
       />
 
+      {/* LoadingButton changes depending on the submitting state of the form and whether TextField is empty or not. */}
       {submitting ? (
         <LoadingButton
           style={{ marginLeft: 20 }}
